requirements.txt
cryptography==42.0.5
Flask==3.0.2
Flask-HTTPAuth==4.8.0
Flask-SQLAlchemy==3.1.1
mysqlclient==2.2.4
requests==2.31.0
PyJWT==2.8.0
python-dotenv==1.0.1